requirements.txt
-i https://pypi.org/simple
streamlit>=1.26.0
langchain==0.0.300
openai==0.28.0
tiktoken==0.5.1
trubrics==1.6.1
streamlit-feedback==0.1.2
youtube-search-python
yt-dlp
openai-whisper
chromadb==0.4.12
openai==0.28.0
python-dotenv==1.0.0